Add a bunch of parameters so Tez actally runs without problems

+++ dhive/scripts/assure_tez.sh
@@ -13,18 +13,34 @@ if [ "$TEZ_COMPILE" = "1" ]; then
     else
         echo "Compiling Tez and copying"
         pushd $TEZ_PAZ
-        mvn clean package -DskipTests=true -Dmaven.javadoc.skip=true || { echo 'Error compiling' ; exit 1; }
+        mvn clean package -pl '!tez-ui' -DskipTests=true -Dmaven.javadoc.skip=true || { echo 'Error compiling' ; exit 1; }
         popd
     fi
 
     cp $TEZ_PAZ/tez-dist/target/tez-$TEZ_VERSION.tar.gz tez.tar.gz || { echo 'Copy failed' ; exit 3; }
+    cp $TEZ_PAZ/tez-dist/target/tez-$TEZ_VERSION.tar.gz tez_up.tar.gz || { echo 'Copy failed' ; exit 3; }
 else
     if [[ -z "${TEZ_PAZ}" ]]; then
-        echo "Downloading tez"
-        wget -nc $TEZ_URL -O tez.tar.gz || true
+        if [ ! -f tez.tar.gz ]; then
+            echo "Downloading tez"
+            wget -nc $TEZ_URL -O tez.tar.gz || true
+
+            # All this is to remove the parent directory from the tar
+            rm -rf tez_temp
+            mkdir tez_temp
+            tar -xvzf tez.tar.gz -C tez_temp
+            pushd tez_temp/*
+            cp share/tez.tar.gz ../../tez_up.tar.gz
+            tar czf ../../tez.tar.gz *
+            popd
+            rm -rf tez_temp
+
+        fi
     else
         cp $TEZ_PAZ/tez-dist/target/tez-$TEZ_VERSION.tar.gz tez.tar.gz || { echo 'Copy failed' ; exit 3; }
+        cp $TEZ_PAZ/tez-dist/target/tez-$TEZ_VERSION.tar.gz tez_up.tar.gz || { echo 'Copy failed' ; exit 3; }
     fi
 fi
 
 cp tez.tar.gz $BASE_CONTAINER_PATH
+cp tez_up.tar.gz $BASE_CONTAINER_PATH
